docker: Generate .dockerignore for BUILDKIT
Generate `_build/Dockerfile.dockerignore` files automatically from all
the `.gitignore` files in the repo. This requires the following
transformations:

1. In `.gitignore` a pattern without a leading or middle `/` applies
also in all subdirectories. In `.dockerignore` such patterns behave
the same as if they started with a `/`. A special `**/` wildcard needs
to be prepended to apply the patters in any (sub) directory.

2. `.gitignore` patterns are relative to the path in which they
reside, while `.dockerignore` is always relative to the docker build
context root. Hence the subdirectory of a `.gitignore` file must be
prepended to the pattern.

3. Inclusion flag `!` must be kept at the beginning of the line
through the above transformations.

With this the docker build context size on my repo shrinks from ~700MB
to ~90MB. This helps speed up the builds as well as conserve disk space.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>